mais algumas chamadas

+++ open/src/main/java/br/com/open/api/open/DTO/PostDTO.java
@@ -1,4 +1,4 @@
 package br.com.open.api.open.DTO;
 
-public record PostDTO(int userId, int id, String title, String body) {
+public record PostDTO(int userId, long id, String title, String body) {
 }
